# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PromptPlace API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/members/{memberId}/prompts": {
      "get": {
        "summary": "회원이 작성한 프롬프트 목록 조회",
        "tags": [
          "Member"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "페이징 커서"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "한 페이지에 가져올 개수 (최대 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "프롬프트 목록 조회 성공"
          }
        }
      }
    },
...
    "/api/prompts/searches": {
      "get": {
        "summary": "프롬프트 검색",
        "description": "키워드, 태그, 모델 등을 기준으로 프롬프트를 검색합니다.",
        "tags": [
          "Prompts"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": true,
            "description": "검색어",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "description": "태그 (여러 개 가능)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "in": "query",
            "name": "model",
            "description": "모델 이름",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "size",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "recent",
                "views",
                "popular"
              ],
              "default": "recent"
            }
          },
          {
            "in": "query",
            "name": "is_free",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "검색 성공"
          },
          "400": {
            "description": "잘못된 요청"
          }
        }
      }
    },
    "/api/prompts/{promptId}/images": {
      "post": {
        "summary": "프롬프트에 이미지 등록",
        "tags": [
          "Prompts"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "promptId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "image_url": {
                    "type": "string"
                  },
                  "order_index": {
                    "type": "integer"
                  }
                },
                "required": [
                  "image_url"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "이미지 매핑 성공"
          },
          "400": {
            "description": "필수 값 누락"
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    },
    "/api/prompts": {
      "post": {
        "summary": "프롬프트 작성",
        "tags": [
          "Prompts"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "prompt",
                  "prompt_result",
                  "description",
                  "price",
                  "tags",
                  "model",
                  "is_free"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "prompt": {
                    "type": "string"
                  },
                  "prompt_result": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "usage_guide": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "model": {
                    "type": "string"
                  },
                  "is_free": {
                    "type": "boolean"
                  },
                  "has_image": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "업로드 성공"
          },
          "400": {
            "description": "필수 필드 누락"
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    },
    "/api/prompts/{promptId}": {
      "patch": {
        "summary": "프롬프트 수정",
        "tags": [
          "Prompts"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "promptId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "prompt": {
                    "type": "string"
                  },
                  "prompt_result": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "usage_guide": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "model": {
                    "type": "string"
                  },
                  "is_free": {
                    "type": "boolean"
                  },
                  "has_image": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 성공"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "프롬프트 없음"
          }
        }
      },
      "delete": {
        "summary": "프롬프트 삭제",
        "tags": [
          "Prompts"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "promptId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 성공"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "프롬프트 없음"
          }
        }
      }
    },
...
  },
...
}